LayoutTests/imported/w3c/web-platform-tests/css/selectors/attribute-selectors/attribute-case/syntax-expected.txt


PASS [foo='BAR'] /* sanity check (valid) */ in standards mode
PASS [foo='BAR'] /* sanity check (valid) */ with querySelector in standards mode
PASS [baz='quux'] /* sanity check (valid) */ in standards mode
PASS [baz='quux'] /* sanity check (valid) */ with querySelector in standards mode
PASS [foo='bar' i] in standards mode
PASS [foo='bar' i] with querySelector in standards mode
PASS [foo='bar' I] in standards mode
PASS [foo='bar' I] with querySelector in standards mode
PASS [foo=bar i] in standards mode
PASS [foo=bar i] with querySelector in standards mode
PASS [foo="bar" i] in standards mode
PASS [foo="bar" i] with querySelector in standards mode
PASS [foo='bar'i] in standards mode
PASS [foo='bar'i] with querySelector in standards mode
PASS [foo='bar'i ] in standards mode
PASS [foo='bar'i ] with querySelector in standards mode
PASS [foo='bar' i ] in standards mode
PASS [foo='bar' i ] with querySelector in standards mode
PASS [foo='bar' /**/ i] in standards mode
PASS [foo='bar' /**/ i] with querySelector in standards mode
PASS [foo='bar' i /**/ ] in standards mode
PASS [foo='bar' i /**/ ] with querySelector in standards mode
PASS [foo='bar'/**/i/**/] in standards mode
PASS [foo='bar'/**/i/**/] with querySelector in standards mode
PASS [foo=bar/**/i] in standards mode
PASS [foo=bar/**/i] with querySelector in standards mode
PASS [foo='bar'	i	] /* \t */ in standards mode
PASS [foo='bar'	i	] /* \t */ with querySelector in standards mode
PASS [foo='bar'
i
] /* \n */ in standards mode
PASS [foo='bar'
i
] /* \n */ with querySelector in standards mode
PASS [foo='bar'\ri\r] /* \r */ in standards mode
PASS [foo='bar'\ri\r] /* \r */ with querySelector in standards mode
PASS [foo='bar' \i] in standards mode
PASS [foo='bar' \i] with querySelector in standards mode
PASS [foo='bar' \69] in standards mode
PASS [foo='bar' \69] with querySelector in standards mode
PASS [foo='bar' \49] in standards mode
PASS [foo='bar' \49] with querySelector in standards mode
PASS [foo~='bar' i] in standards mode
PASS [foo~='bar' i] with querySelector in standards mode
PASS [foo^='bar' i] in standards mode
PASS [foo^='bar' i] with querySelector in standards mode
PASS [foo$='bar' i] in standards mode
PASS [foo$='bar' i] with querySelector in standards mode
PASS [foo*='bar' i] in standards mode
PASS [foo*='bar' i] with querySelector in standards mode
PASS [foo|='bar' i] in standards mode
PASS [foo|='bar' i] with querySelector in standards mode
PASS [|foo='bar' i] in standards mode
PASS [|foo='bar' i] with querySelector in standards mode
PASS [*|foo='bar' i] in standards mode
PASS [*|foo='bar' i] with querySelector in standards mode
FAIL [baz='quux' s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux' S] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' S] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz=quux s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz=quux s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz="quux" s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz="quux" s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux's] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux's] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux's ] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux's ] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux' s ] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' s ] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux' /**/ s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' /**/ s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux' s /**/ ] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' s /**/ ] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux'/**/s/**/] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'/**/s/**/] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz=quux/**/s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz=quux/**/s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux'	s	] /* \t */ in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'	s	] /* \t */ with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux'
s
] /* \n */ in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'
s
] /* \n */ with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux'\rs\r] /* \r */ in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'\rs\r] /* \r */ with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux' \s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' \s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux' \73] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' \73] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz='quux' \53] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' \53] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz~='quux' s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz~='quux' s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz^='quux' s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz^='quux' s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz$='quux' s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz$='quux' s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz*='quux' s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz*='quux' s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [baz|='quux' s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz|='quux' s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [|baz='quux' s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [|baz='quux' s] with querySelector in standards mode The string did not match the expected pattern.
FAIL [*|baz='quux' s] in standards mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [*|baz='quux' s] with querySelector in standards mode The string did not match the expected pattern.
PASS [foo[ /* sanity check (invalid) */ in standards mode
PASS [foo[ /* sanity check (invalid) */ with querySelector in standards mode
PASS [foo='bar' i i] in standards mode
PASS [foo='bar' i i] with querySelector in standards mode
PASS [foo i ='bar'] in standards mode
PASS [foo i ='bar'] with querySelector in standards mode
PASS [foo= i 'bar'] in standards mode
PASS [foo= i 'bar'] with querySelector in standards mode
PASS [i foo='bar'] in standards mode
PASS [i foo='bar'] with querySelector in standards mode
PASS [foo='bar' i\0] /* \0 */ in standards mode
PASS [foo='bar' i\0] /* \0 */ with querySelector in standards mode
PASS [foo='bar' İ] in standards mode
PASS [foo='bar' İ] with querySelector in standards mode
PASS [foo='bar' ı] in standards mode
PASS [foo='bar' ı] with querySelector in standards mode
PASS [foo='bar' ii] in standards mode
PASS [foo='bar' ii] with querySelector in standards mode
PASS [foo='bar' ij] in standards mode
PASS [foo='bar' ij] with querySelector in standards mode
PASS [foo='bar' j] in standards mode
PASS [foo='bar' j] with querySelector in standards mode
PASS [foo='bar' \\i] in standards mode
PASS [foo='bar' \\i] with querySelector in standards mode
PASS [foo='bar' \\69] in standards mode
PASS [foo='bar' \\69] with querySelector in standards mode
PASS [foo='bar' i()] in standards mode
PASS [foo='bar' i()] with querySelector in standards mode
PASS [foo='bar' i ()] in standards mode
PASS [foo='bar' i ()] with querySelector in standards mode
PASS [foo='bar' () i] in standards mode
PASS [foo='bar' () i] with querySelector in standards mode
PASS [foo='bar' (i)] in standards mode
PASS [foo='bar' (i)] with querySelector in standards mode
PASS [foo='bar' i []] in standards mode
PASS [foo='bar' i []] with querySelector in standards mode
PASS [foo='bar' [] i] in standards mode
PASS [foo='bar' [] i] with querySelector in standards mode
PASS [foo='bar' [i]] in standards mode
PASS [foo='bar' [i]] with querySelector in standards mode
PASS [foo='bar' i {}] in standards mode
PASS [foo='bar' i {}] with querySelector in standards mode
PASS [foo='bar' {} i] in standards mode
PASS [foo='bar' {} i] with querySelector in standards mode
PASS [foo='bar' {i}] in standards mode
PASS [foo='bar' {i}] with querySelector in standards mode
PASS [foo='bar' 1i] in standards mode
PASS [foo='bar' 1i] with querySelector in standards mode
PASS [foo='bar' 1] in standards mode
PASS [foo='bar' 1] with querySelector in standards mode
PASS [foo='bar' 'i'] in standards mode
PASS [foo='bar' 'i'] with querySelector in standards mode
PASS [foo='bar' url(i)] in standards mode
PASS [foo='bar' url(i)] with querySelector in standards mode
PASS [foo='bar' ,i] in standards mode
PASS [foo='bar' ,i] with querySelector in standards mode
PASS [foo='bar' i,] in standards mode
PASS [foo='bar' i,] with querySelector in standards mode
PASS [foo='bar']i in standards mode
PASS [foo='bar']i with querySelector in standards mode
PASS [foo='bar' |i] in standards mode
PASS [foo='bar' |i] with querySelector in standards mode
PASS [foo='bar' \|i] in standards mode
PASS [foo='bar' \|i] with querySelector in standards mode
PASS [foo='bar' *|i] in standards mode
PASS [foo='bar' *|i] with querySelector in standards mode
PASS [foo='bar' \*|i] in standards mode
PASS [foo='bar' \*|i] with querySelector in standards mode
PASS [foo='bar' *] in standards mode
PASS [foo='bar' *] with querySelector in standards mode
PASS [foo='bar' \*] in standards mode
PASS [foo='bar' \*] with querySelector in standards mode
PASS [foo i] in standards mode
PASS [foo i] with querySelector in standards mode
PASS [foo/**/i] in standards mode
PASS [foo/**/i] with querySelector in standards mode
PASS [foo='BAR'] /* sanity check (valid) */ in quirks mode
PASS [foo='BAR'] /* sanity check (valid) */ with querySelector in quirks mode
PASS [baz='quux'] /* sanity check (valid) */ in quirks mode
PASS [baz='quux'] /* sanity check (valid) */ with querySelector in quirks mode
PASS [foo='bar' i] in quirks mode
PASS [foo='bar' i] with querySelector in quirks mode
PASS [foo='bar' I] in quirks mode
PASS [foo='bar' I] with querySelector in quirks mode
PASS [foo=bar i] in quirks mode
PASS [foo=bar i] with querySelector in quirks mode
PASS [foo="bar" i] in quirks mode
PASS [foo="bar" i] with querySelector in quirks mode
PASS [foo='bar'i] in quirks mode
PASS [foo='bar'i] with querySelector in quirks mode
PASS [foo='bar'i ] in quirks mode
PASS [foo='bar'i ] with querySelector in quirks mode
PASS [foo='bar' i ] in quirks mode
PASS [foo='bar' i ] with querySelector in quirks mode
PASS [foo='bar' /**/ i] in quirks mode
PASS [foo='bar' /**/ i] with querySelector in quirks mode
PASS [foo='bar' i /**/ ] in quirks mode
PASS [foo='bar' i /**/ ] with querySelector in quirks mode
PASS [foo='bar'/**/i/**/] in quirks mode
PASS [foo='bar'/**/i/**/] with querySelector in quirks mode
PASS [foo=bar/**/i] in quirks mode
PASS [foo=bar/**/i] with querySelector in quirks mode
PASS [foo='bar'	i	] /* \t */ in quirks mode
PASS [foo='bar'	i	] /* \t */ with querySelector in quirks mode
PASS [foo='bar'
i
] /* \n */ in quirks mode
PASS [foo='bar'
i
] /* \n */ with querySelector in quirks mode
PASS [foo='bar'\ri\r] /* \r */ in quirks mode
PASS [foo='bar'\ri\r] /* \r */ with querySelector in quirks mode
PASS [foo='bar' \i] in quirks mode
PASS [foo='bar' \i] with querySelector in quirks mode
PASS [foo='bar' \69] in quirks mode
PASS [foo='bar' \69] with querySelector in quirks mode
PASS [foo='bar' \49] in quirks mode
PASS [foo='bar' \49] with querySelector in quirks mode
PASS [foo~='bar' i] in quirks mode
PASS [foo~='bar' i] with querySelector in quirks mode
PASS [foo^='bar' i] in quirks mode
PASS [foo^='bar' i] with querySelector in quirks mode
PASS [foo$='bar' i] in quirks mode
PASS [foo$='bar' i] with querySelector in quirks mode
PASS [foo*='bar' i] in quirks mode
PASS [foo*='bar' i] with querySelector in quirks mode
PASS [foo|='bar' i] in quirks mode
PASS [foo|='bar' i] with querySelector in quirks mode
PASS [|foo='bar' i] in quirks mode
PASS [|foo='bar' i] with querySelector in quirks mode
PASS [*|foo='bar' i] in quirks mode
PASS [*|foo='bar' i] with querySelector in quirks mode
FAIL [baz='quux' s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux' S] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' S] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz=quux s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz=quux s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz="quux" s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz="quux" s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux's] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux's] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux's ] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux's ] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux' s ] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' s ] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux' /**/ s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' /**/ s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux' s /**/ ] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' s /**/ ] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux'/**/s/**/] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'/**/s/**/] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz=quux/**/s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz=quux/**/s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux'	s	] /* \t */ in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'	s	] /* \t */ with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux'
s
] /* \n */ in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'
s
] /* \n */ with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux'\rs\r] /* \r */ in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'\rs\r] /* \r */ with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux' \s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' \s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux' \73] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' \73] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz='quux' \53] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' \53] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz~='quux' s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz~='quux' s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz^='quux' s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz^='quux' s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz$='quux' s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz$='quux' s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz*='quux' s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz*='quux' s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [baz|='quux' s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz|='quux' s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [|baz='quux' s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [|baz='quux' s] with querySelector in quirks mode The string did not match the expected pattern.
FAIL [*|baz='quux' s] in quirks mode assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [*|baz='quux' s] with querySelector in quirks mode The string did not match the expected pattern.
PASS [foo[ /* sanity check (invalid) */ in quirks mode
PASS [foo[ /* sanity check (invalid) */ with querySelector in quirks mode
PASS [foo='bar' i i] in quirks mode
PASS [foo='bar' i i] with querySelector in quirks mode
PASS [foo i ='bar'] in quirks mode
PASS [foo i ='bar'] with querySelector in quirks mode
PASS [foo= i 'bar'] in quirks mode
PASS [foo= i 'bar'] with querySelector in quirks mode
PASS [i foo='bar'] in quirks mode
PASS [i foo='bar'] with querySelector in quirks mode
PASS [foo='bar' i\0] /* \0 */ in quirks mode
PASS [foo='bar' i\0] /* \0 */ with querySelector in quirks mode
PASS [foo='bar' İ] in quirks mode
PASS [foo='bar' İ] with querySelector in quirks mode
PASS [foo='bar' ı] in quirks mode
PASS [foo='bar' ı] with querySelector in quirks mode
PASS [foo='bar' ii] in quirks mode
PASS [foo='bar' ii] with querySelector in quirks mode
PASS [foo='bar' ij] in quirks mode
PASS [foo='bar' ij] with querySelector in quirks mode
PASS [foo='bar' j] in quirks mode
PASS [foo='bar' j] with querySelector in quirks mode
PASS [foo='bar' \\i] in quirks mode
PASS [foo='bar' \\i] with querySelector in quirks mode
PASS [foo='bar' \\69] in quirks mode
PASS [foo='bar' \\69] with querySelector in quirks mode
PASS [foo='bar' i()] in quirks mode
PASS [foo='bar' i()] with querySelector in quirks mode
PASS [foo='bar' i ()] in quirks mode
PASS [foo='bar' i ()] with querySelector in quirks mode
PASS [foo='bar' () i] in quirks mode
PASS [foo='bar' () i] with querySelector in quirks mode
PASS [foo='bar' (i)] in quirks mode
PASS [foo='bar' (i)] with querySelector in quirks mode
PASS [foo='bar' i []] in quirks mode
PASS [foo='bar' i []] with querySelector in quirks mode
PASS [foo='bar' [] i] in quirks mode
PASS [foo='bar' [] i] with querySelector in quirks mode
PASS [foo='bar' [i]] in quirks mode
PASS [foo='bar' [i]] with querySelector in quirks mode
PASS [foo='bar' i {}] in quirks mode
PASS [foo='bar' i {}] with querySelector in quirks mode
PASS [foo='bar' {} i] in quirks mode
PASS [foo='bar' {} i] with querySelector in quirks mode
PASS [foo='bar' {i}] in quirks mode
PASS [foo='bar' {i}] with querySelector in quirks mode
PASS [foo='bar' 1i] in quirks mode
PASS [foo='bar' 1i] with querySelector in quirks mode
PASS [foo='bar' 1] in quirks mode
PASS [foo='bar' 1] with querySelector in quirks mode
PASS [foo='bar' 'i'] in quirks mode
PASS [foo='bar' 'i'] with querySelector in quirks mode
PASS [foo='bar' url(i)] in quirks mode
PASS [foo='bar' url(i)] with querySelector in quirks mode
PASS [foo='bar' ,i] in quirks mode
PASS [foo='bar' ,i] with querySelector in quirks mode
PASS [foo='bar' i,] in quirks mode
PASS [foo='bar' i,] with querySelector in quirks mode
PASS [foo='bar']i in quirks mode
PASS [foo='bar']i with querySelector in quirks mode
PASS [foo='bar' |i] in quirks mode
PASS [foo='bar' |i] with querySelector in quirks mode
PASS [foo='bar' \|i] in quirks mode
PASS [foo='bar' \|i] with querySelector in quirks mode
PASS [foo='bar' *|i] in quirks mode
PASS [foo='bar' *|i] with querySelector in quirks mode
PASS [foo='bar' \*|i] in quirks mode
PASS [foo='bar' \*|i] with querySelector in quirks mode
PASS [foo='bar' *] in quirks mode
PASS [foo='bar' *] with querySelector in quirks mode
PASS [foo='bar' \*] in quirks mode
PASS [foo='bar' \*] with querySelector in quirks mode
PASS [foo i] in quirks mode
PASS [foo i] with querySelector in quirks mode
PASS [foo/**/i] in quirks mode
PASS [foo/**/i] with querySelector in quirks mode
PASS [foo='BAR'] /* sanity check (valid) */ in XML
PASS [foo='BAR'] /* sanity check (valid) */ with querySelector in XML
PASS [baz='quux'] /* sanity check (valid) */ in XML
PASS [baz='quux'] /* sanity check (valid) */ with querySelector in XML
PASS [foo='bar' i] in XML
PASS [foo='bar' i] with querySelector in XML
PASS [foo='bar' I] in XML
PASS [foo='bar' I] with querySelector in XML
PASS [foo=bar i] in XML
PASS [foo=bar i] with querySelector in XML
PASS [foo="bar" i] in XML
PASS [foo="bar" i] with querySelector in XML
PASS [foo='bar'i] in XML
PASS [foo='bar'i] with querySelector in XML
PASS [foo='bar'i ] in XML
PASS [foo='bar'i ] with querySelector in XML
PASS [foo='bar' i ] in XML
PASS [foo='bar' i ] with querySelector in XML
PASS [foo='bar' /**/ i] in XML
PASS [foo='bar' /**/ i] with querySelector in XML
PASS [foo='bar' i /**/ ] in XML
PASS [foo='bar' i /**/ ] with querySelector in XML
PASS [foo='bar'/**/i/**/] in XML
PASS [foo='bar'/**/i/**/] with querySelector in XML
PASS [foo=bar/**/i] in XML
PASS [foo=bar/**/i] with querySelector in XML
PASS [foo='bar'	i	] /* \t */ in XML
PASS [foo='bar'	i	] /* \t */ with querySelector in XML
PASS [foo='bar'
i
] /* \n */ in XML
PASS [foo='bar'
i
] /* \n */ with querySelector in XML
PASS [foo='bar'\ri\r] /* \r */ in XML
PASS [foo='bar'\ri\r] /* \r */ with querySelector in XML
PASS [foo='bar' \i] in XML
PASS [foo='bar' \i] with querySelector in XML
PASS [foo='bar' \69] in XML
PASS [foo='bar' \69] with querySelector in XML
PASS [foo='bar' \49] in XML
PASS [foo='bar' \49] with querySelector in XML
PASS [foo~='bar' i] in XML
PASS [foo~='bar' i] with querySelector in XML
PASS [foo^='bar' i] in XML
PASS [foo^='bar' i] with querySelector in XML
PASS [foo$='bar' i] in XML
PASS [foo$='bar' i] with querySelector in XML
PASS [foo*='bar' i] in XML
PASS [foo*='bar' i] with querySelector in XML
PASS [foo|='bar' i] in XML
PASS [foo|='bar' i] with querySelector in XML
PASS [|foo='bar' i] in XML
PASS [|foo='bar' i] with querySelector in XML
PASS [*|foo='bar' i] in XML
PASS [*|foo='bar' i] with querySelector in XML
FAIL [baz='quux' s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux' S] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' S] with querySelector in XML The string did not match the expected pattern.
FAIL [baz=quux s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz=quux s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz="quux" s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz="quux" s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux's] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux's] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux's ] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux's ] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux' s ] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' s ] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux' /**/ s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' /**/ s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux' s /**/ ] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' s /**/ ] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux'/**/s/**/] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'/**/s/**/] with querySelector in XML The string did not match the expected pattern.
FAIL [baz=quux/**/s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz=quux/**/s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux'	s	] /* \t */ in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'	s	] /* \t */ with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux'
s
] /* \n */ in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'
s
] /* \n */ with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux'\rs\r] /* \r */ in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux'\rs\r] /* \r */ with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux' \s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' \s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux' \73] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' \73] with querySelector in XML The string did not match the expected pattern.
FAIL [baz='quux' \53] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz='quux' \53] with querySelector in XML The string did not match the expected pattern.
FAIL [baz~='quux' s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz~='quux' s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz^='quux' s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz^='quux' s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz$='quux' s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz$='quux' s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz*='quux' s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz*='quux' s] with querySelector in XML The string did not match the expected pattern.
FAIL [baz|='quux' s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [baz|='quux' s] with querySelector in XML The string did not match the expected pattern.
FAIL [|baz='quux' s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [|baz='quux' s] with querySelector in XML The string did not match the expected pattern.
FAIL [*|baz='quux' s] in XML assert_equals: valid rule didn't parse into CSSOM expected 1 but got 0
FAIL [*|baz='quux' s] with querySelector in XML The string did not match the expected pattern.
PASS [foo[ /* sanity check (invalid) */ in XML
PASS [foo[ /* sanity check (invalid) */ with querySelector in XML
PASS [foo='bar' i i] in XML
PASS [foo='bar' i i] with querySelector in XML
PASS [foo i ='bar'] in XML
PASS [foo i ='bar'] with querySelector in XML
PASS [foo= i 'bar'] in XML
PASS [foo= i 'bar'] with querySelector in XML
PASS [i foo='bar'] in XML
PASS [i foo='bar'] with querySelector in XML
PASS [foo='bar' i\0] /* \0 */ in XML
PASS [foo='bar' i\0] /* \0 */ with querySelector in XML
PASS [foo='bar' İ] in XML
PASS [foo='bar' İ] with querySelector in XML
PASS [foo='bar' ı] in XML
PASS [foo='bar' ı] with querySelector in XML
PASS [foo='bar' ii] in XML
PASS [foo='bar' ii] with querySelector in XML
PASS [foo='bar' ij] in XML
PASS [foo='bar' ij] with querySelector in XML
PASS [foo='bar' j] in XML
PASS [foo='bar' j] with querySelector in XML
PASS [foo='bar' \\i] in XML
PASS [foo='bar' \\i] with querySelector in XML
PASS [foo='bar' \\69] in XML
PASS [foo='bar' \\69] with querySelector in XML
PASS [foo='bar' i()] in XML
PASS [foo='bar' i()] with querySelector in XML
PASS [foo='bar' i ()] in XML
PASS [foo='bar' i ()] with querySelector in XML
PASS [foo='bar' () i] in XML
PASS [foo='bar' () i] with querySelector in XML
PASS [foo='bar' (i)] in XML
PASS [foo='bar' (i)] with querySelector in XML
PASS [foo='bar' i []] in XML
PASS [foo='bar' i []] with querySelector in XML
PASS [foo='bar' [] i] in XML
PASS [foo='bar' [] i] with querySelector in XML
PASS [foo='bar' [i]] in XML
PASS [foo='bar' [i]] with querySelector in XML
PASS [foo='bar' i {}] in XML
PASS [foo='bar' i {}] with querySelector in XML
PASS [foo='bar' {} i] in XML
PASS [foo='bar' {} i] with querySelector in XML
PASS [foo='bar' {i}] in XML
PASS [foo='bar' {i}] with querySelector in XML
PASS [foo='bar' 1i] in XML
PASS [foo='bar' 1i] with querySelector in XML
PASS [foo='bar' 1] in XML
PASS [foo='bar' 1] with querySelector in XML
PASS [foo='bar' 'i'] in XML
PASS [foo='bar' 'i'] with querySelector in XML
PASS [foo='bar' url(i)] in XML
PASS [foo='bar' url(i)] with querySelector in XML
PASS [foo='bar' ,i] in XML
PASS [foo='bar' ,i] with querySelector in XML
PASS [foo='bar' i,] in XML
PASS [foo='bar' i,] with querySelector in XML
PASS [foo='bar']i in XML
PASS [foo='bar']i with querySelector in XML
PASS [foo='bar' |i] in XML
PASS [foo='bar' |i] with querySelector in XML
PASS [foo='bar' \|i] in XML
PASS [foo='bar' \|i] with querySelector in XML
PASS [foo='bar' *|i] in XML
PASS [foo='bar' *|i] with querySelector in XML
PASS [foo='bar' \*|i] in XML
PASS [foo='bar' \*|i] with querySelector in XML
PASS [foo='bar' *] in XML
PASS [foo='bar' *] with querySelector in XML
PASS [foo='bar' \*] in XML
PASS [foo='bar' \*] with querySelector in XML
PASS [foo i] in XML
PASS [foo i] with querySelector in XML
PASS [foo/**/i] in XML
PASS [foo/**/i] with querySelector in XML
